pyproject.toml
[project]
name = "ower"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "arrio", email = ""},
]
dependencies = [
    "requests>=2.32.3",
    "pyaudio>=0.2.14",
    "openai>=1.31.0",
    "paho-mqtt>=2.1.0",
    "gtts>=2.5.1",
    "pygame>=2.5.2",
    "opencv-python>=4.10.0.82",
    "pillow>=10.3.0",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
